+++ Changes.md
@@ -9,6 +9,7 @@
 * Only permit parameters but don't require them. Requiring them raises an exception if they're missing from the form
   submission. But in cases where other parameters are sent as well and the form object may be optional this would raise
   an exception that would be undesired.
+* Allow all forms to be signed by default.
 
 ## 0.1.2